# modules/marketplace/OPENAPI.yaml
openapi: 3.1.0
info:
  title: All-in-One Marketplace API
  version: 0.2.0
servers:
  - url: /api/v1/marketplace
security:
  - bearerAuth: []
paths:
...
  /valley/orders/{order_id}/support:
    post:
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupportCaseRequest'
      responses:
        '201':
          description: Support or dispute case created
        '403':
          description: Order does not belong to the authenticated customer
        '409':
          description: Order is not eligible for support yet
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    SupportCaseRequest:
      type: object
      required: [kind, message, idempotency_key]
      properties:
        kind:
          type: string
          enum: [support, dispute]
        subject:
          type: string
        message:
          type: string
        desired_resolution:
          type: string
        idempotency_key:
          type: string